Modified A* Pathfinding Algorithm to support multi-tile characters.
Using the collision box of the Start character, we now can find the tile locations relative to the main tile location (where the character is present in the world), and check those list of tiles when computing whether or not a neighbor is traversable. This incidentally caused no path to be found if the Target character is standing against a wall or in a space that the Start character cannot reach. In order to solve this problem, we simply search for the Node in the closed set with the lowest f-score and choose to go there.